banking-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlmodel==0.0.21
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.3.2
httpx==0.27.2
